Cardano.Crypto.Signing.Tag

Documentation

data SignTag

To protect against replay attacks (i.e. when an attacker intercepts a signed piece of data and later sends it again), we add a tag to all data that we sign. This ensures that even if some bytestring can be deserialized into two different types of messages (A and B), the attacker can't take message A and send it as message B.

We also automatically add the network tag (protocolMagic) whenever it makes sense, to ensure that things intended for testnet won't work for mainnet.

Constructors

SignForTestingOnly
Anything (to be used for testing only)

SignTx
Tx: TxSigData

SignRedeemTx
Redeem tx: TxSigData

SignVssCert
Vss certificate: (VssVerificationKey, EpochNumber)

SignUSProposal
Update proposal: UpdateProposalToSign

SignCommitment
Commitment: (EpochNumber, Commitment)

SignUSVote
US proposal vote: (UpId, Bool)

SignBlock VerificationKey
Block header: ToSign

This constructor takes the VerificationKey of the delegation certificate issuer, which is prepended to the signature as part of the sign tag

SignCertificate
Certificate: Certificate

signTag ∷ ProtocolMagicId → SignTag → ByteString

Get magic bytes corresponding to a SignTag. Guaranteed to be different (and begin with a different byte) for different tags.

signTagDecoded ∷ Annotated ProtocolMagicId ByteString → SignTag → ByteString

Get magic bytes corresponding to a SignTag, taking ProtocolMagic bytes from the annotation