Cardano.Crypto.Hashing

Description

Hashing capabilities.

AbstractHash type supporting different hash algorithms

data AbstractHash algo a

Hash wrapper with phantom type for more type-safety

Made abstract in order to support different algorithms

class HashAlgorithm a

Minimal complete definition

hashBlockSize, hashDigestSize, hashInternalContextSize, hashInternalInit, hashInternalUpdate, hashInternalFinalize

Hashing

abstractHash :: (HashAlgorithm algo, EncCBOR a) => a -> AbstractHash algo a

Hash the EncCBOR-serialised version of a value Once this is no longer used outside this module it should be made private.

unsafeAbstractHash :: HashAlgorithm algo => LByteString -> AbstractHash algo a

Hash a lazy LByteString

You can choose the phantom type, hence the "unsafe".

Conversion

abstractHashFromDigest :: Digest algo -> AbstractHash algo a

Make an AbstractHash from a Digest for the same HashAlgorithm.

abstractHashFromBytes :: forall algo a. HashAlgorithm algo => ByteString -> Maybe (AbstractHash algo a)

Make an AbstractHash from the bytes representation of the hash. It will fail if given the wrong number of bytes for the choice of HashAlgorithm.

unsafeAbstractHashFromBytes :: ByteString -> AbstractHash algo a

Like abstractHashFromDigestBytes but the number of bytes provided must be correct for the choice of HashAlgorithm.

abstractHashToBytes :: AbstractHash algo a -> ByteString

The bytes representation of the hash value.

unsafeAbstractHashFromShort :: ShortByteString -> AbstractHash algo a

The ShortByteString representation of the hash value.

abstractHashToShort :: AbstractHash algo a -> ShortByteString

The ShortByteString representation of the hash value.

Parsing and printing

decodeAbstractHash :: HashAlgorithm algo => Text -> Either Text (AbstractHash algo a)

Parses given hash in base16 form.

Standard Hash type using Blake2b 256

type Hash = AbstractHash Blake2b_256

The type of our commonly used hash, Blake2b 256

Hashing

hash :: EncCBOR a => a -> Hash a

Deprecated: Use serializeCborHash or hash the annotation instead.

The hash of a value, serialised via EncCBOR.

hashDecoded :: Decoded t => t -> Hash (BaseType t)

The hash of a value's annotation

hashRaw :: ByteString -> Hash Raw

Hash a bytestring

serializeCborHash :: EncCBOR a => a -> Hash a

The hash of a value, serialised via EncCBOR.

Conversion

hashFromBytes :: ByteString -> Maybe (Hash a)

Make a hash from it bytes representation. It must be a 32-byte bytestring. The size is checked.

unsafeHashFromBytes :: ByteString -> Hash a

Make a hash from a 32-byte bytestring. It must be exactly 32 bytes.

hashToBytes :: AbstractHash algo a -> ByteString

The bytes representation of the hash value.

Parsing and printing

decodeHash :: Text -> Either Text (Hash a)

Parses given hash in base16 form.

hashHexF :: Format r (AbstractHash algo a -> r)

Specialized formatter for Hash.

mediumHashF :: Format r (AbstractHash algo a -> r)

Smart formatter for Hash to show only first 16 characters of Hash.

shortHashF :: Format r (AbstractHash algo a -> r)

Smart formatter for Hash to show only first 8 characters of Hash.