Cardano.Crypto.Hashing

Description

Hashing capabilities.

AbstractHash

newtype AbstractHash algo a

Hash wrapper with phantom type for more type-safety

Made abstract in order to support different algorithms

Constructors

AbstractHash (Digest algo)

decodeAbstractHash ∷ HashAlgorithm algo ⇒ Text → Either Text (AbstractHash algo a)

Parses given hash in base16 form.

decodeHash ∷ Text → Either Text (Hash a)

Parses given hash in base16 form.

abstractHash ∷ (HashAlgorithm algo, ToCBOR a) ⇒ a → AbstractHash algo a

Hash the ToCBOR-serialised version of a value

unsafeAbstractHash ∷ HashAlgorithm algo ⇒ LByteString → AbstractHash algo anything

Make an AbstractHash from a lazy ByteString

You can choose the phantom type, hence the "unsafe"

Common Hash

type Hash = AbstractHash Blake2b_256

Type alias for commonly used hash

hashHexF ∷ Format r (AbstractHash algo a → r)

Specialized formatter for Hash.

mediumHashF ∷ Format r (AbstractHash algo a → r)

Smart formatter for Hash to show only first 16 characters of Hash.

shortHashF ∷ Format r (AbstractHash algo a → r)

Smart formatter for Hash to show only first 8 characters of Hash.

hash ∷ ToCBOR a ⇒ a → Hash a

Short version of unsafeHash.

hashDecoded ∷ Decoded t ⇒ t → Hash (BaseType t)

Hashes the annotation

hashRaw ∷ ByteString → Hash Raw

Raw constructor application.

Utility

class HashAlgorithm a

Class representing hashing algorithms.

The interface presented here is update in place and lowlevel. the Hash module takes care of hidding the mutable interface properly.

Minimal complete definition

hashBlockSize, hashDigestSize, hashInternalContextSize, hashInternalInit, hashInternalUpdate, hashInternalFinalize

hashDigestSize' ∷ ∀ algo. HashAlgorithm algo ⇒ Int