Cardano.Crypto.Raw

Documentation

newtype Raw

A wrapper over ByteString for signalling that a bytestring should be processed as a sequence of bytes, not as a separate entity. It's used in crypto and binary code.

Constructors

Raw ByteString