Cardano.Crypto.Signing

Documentation

newtype VerificationKey

Wrapper around XPub.

Constructors

VerificationKey

Fields

unVerificationKey ∷ XPub

formatFullVerificationKey ∷ VerificationKey → Builder

Builder for VerificationKey to show it in base64 encoded form.

fullVerificationKeyF ∷ Format r (VerificationKey → r)

Formatter for VerificationKey to show it in base64.

fullVerificationKeyHexF ∷ Format r (VerificationKey → r)

Formatter for VerificationKey to show it in hex.

shortVerificationKeyHexF ∷ Format r (VerificationKey → r)

Formatter for VerificationKey to show it in hex, but only first 8 chars.

parseFullVerificationKey ∷ Text → Either VerificationKeyParseError VerificationKey

Parse VerificationKey from base64 encoded string

data SignTag

To protect against replay attacks (i.e. when an attacker intercepts a signed piece of data and later sends it again), we add a tag to all data that we sign. This ensures that even if some bytestring can be deserialized into two different types of messages (A and B), the attacker can't take message A and send it as message B.

We also automatically add the network tag (protocolMagic) whenever it makes sense, to ensure that things intended for testnet won't work for mainnet.

Constructors

SignForTestingOnly
Anything (to be used for testing only)

SignTx
Tx: TxSigData

SignRedeemTx
Redeem tx: TxSigData

SignVssCert
Vss certificate: (VssVerificationKey, EpochNumber)

SignUSProposal
Update proposal: UpdateProposalToSign

SignCommitment
Commitment: (EpochNumber, Commitment)

SignUSVote
US proposal vote: (UpId, Bool)

SignBlock VerificationKey
Block header: ToSign

This constructor takes the VerificationKey of the delegation certificate issuer, which is prepended to the signature as part of the sign tag

SignCertificate
Certificate: Certificate

signTag ∷ ProtocolMagicId → SignTag → ByteString

Get magic bytes corresponding to a SignTag. Guaranteed to be different (and begin with a different byte) for different tags.

signTagDecoded ∷ Annotated ProtocolMagicId ByteString → SignTag → ByteString

Get magic bytes corresponding to a SignTag, taking ProtocolMagic bytes from the annotation

newtype SigningKey

Wrapper around XPrv.

Constructors

SigningKey

Fields

unSigningKey ∷ XPrv

toVerification ∷ SigningKey → VerificationKey

Generate a verification key from a signing key. Fast (it just drops some bytes off the signing key).

encCBORXPrv ∷ XPrv → Encoding

decCBORXPrv ∷ Decoder s XPrv

toCBORXPrv ∷ XPrv → Encoding

fromCBORXPrv ∷ Decoder s XPrv

Signature

newtype Signature a

Wrapper around XSignature

Constructors

Signature XSignature

encCBORXSignature ∷ XSignature → Encoding

decCBORXSignature ∷ Decoder s XSignature

fullSignatureHexF ∷ Format r (Signature a → r)

Formatter for Signature to show it in hex.

parseFullSignature ∷ Text → Either SignatureParseError (Signature a)

Parse Signature from base16 encoded string.

Signing

sign

Arguments

∷ EncCBOR a

⇒ ProtocolMagicId

→ SignTag
See docs for SignTag

→ SigningKey

→ a

→ Signature a

Encode something with EncCBOR and sign it

signEncoded ∷ ProtocolMagicId → SignTag → SigningKey → Encoding → Signature a

Like sign but without the EncCBOR constraint

signRaw

Arguments

∷ ProtocolMagicId

→ Maybe SignTag
See docs for SignTag. Unlike in sign, we allow no tag to be provided just in case you need to sign exactly the bytestring you provided.

→ SigningKey

→ ByteString

→ Signature Raw

Sign a Raw bytestring

safeSign ∷ EncCBOR a ⇒ ProtocolMagicId → SignTag → SafeSigner → a → Signature a

safeSignRaw ∷ ProtocolMagicId → Maybe SignTag → SafeSigner → ByteString → Signature Raw

Verification

verifySignature ∷ (a → Encoding) → ProtocolMagicId → SignTag → VerificationKey → a → Signature a → Bool

Verify a signature

verifySignatureDecoded ∷ Decoded t ⇒ Annotated ProtocolMagicId ByteString → SignTag → VerificationKey → t → Signature (BaseType t) → Bool

Verify a signature

verifySignatureRaw ∷ VerificationKey → ByteString → Signature Raw → Bool

Verify Raw signature

module Cardano.Crypto.Signing.Safe

module Cardano.Crypto.Signing.Redeem

keyGen ∷ MonadRandom m ⇒ m (VerificationKey, SigningKey)

Generate a key pair. It's recommended to run it with runSecureRandom from Cardano.Crypto.Random because the OpenSSL generator is probably safer than the default IO generator.

deterministicKeyGen ∷ ByteString → (VerificationKey, SigningKey)

Create key pair deterministically from 32 bytes.