Cardano.Crypto.Signing.Signature

Signature

newtype Signature a

Wrapper around XSignature

Constructors

Signature XSignature

toCBORXSignature ∷ XSignature → Encoding

fromCBORXSignature ∷ Decoder s XSignature

fullSignatureHexF ∷ Format r (Signature a → r)

Formatter for Signature to show it in hex.

parseFullSignature ∷ Text → Either SignatureParseError (Signature a)

Parse Signature from base16 encoded string.

Signing

sign

Arguments

∷ ToCBOR a

⇒ ProtocolMagicId

→ SignTag
See docs for SignTag

→ SigningKey

→ a

→ Signature a

Encode something with ToCBOR and sign it

signEncoded ∷ ProtocolMagicId → SignTag → SigningKey → Encoding → Signature a

Like sign but without the ToCBOR constraint

signRaw

Arguments

∷ ProtocolMagicId

→ Maybe SignTag
See docs for SignTag. Unlike in sign, we allow no tag to be provided just in case you need to sign exactly the bytestring you provided.

→ SigningKey

→ ByteString

→ Signature Raw

Sign a Raw bytestring

safeSign ∷ ToCBOR a ⇒ ProtocolMagicId → SignTag → SafeSigner → a → Signature a

safeSignRaw ∷ ProtocolMagicId → Maybe SignTag → SafeSigner → ByteString → Signature Raw

Verification

verifySignature ∷ (a → Encoding) → ProtocolMagicId → SignTag → VerificationKey → a → Signature a → Bool

Verify a signature

verifySignatureDecoded ∷ Decoded t ⇒ Annotated ProtocolMagicId ByteString → SignTag → VerificationKey → t → Signature (BaseType t) → Bool

Verify a signature

verifySignatureRaw ∷ VerificationKey → ByteString → Signature Raw → Bool

Verify Raw signature